Latest updates

+++ src/Library.Amazon/S3StorageManager.cs
@@ -10,12 +10,13 @@ using Amazon.S3;
 using Amazon.S3.Model;
 using Amazon.S3.Util;
 using Library.Amazon.Resources;
-using Library.Storage;
+using Library.Platform.Storage;
 
 namespace Library.Amazon
 {
     public class S3StorageManager : IStorageManager
     {
+        private static readonly SemaphoreSlim _createSemaphore = new SemaphoreSlim(1, 1);       // TODO: Shouldn't this be: new SemaphoreSlim(0, 1);? What does initialCount mean?
         private readonly IAmazonS3 _client;
 
         public S3StorageManager(IAmazonS3 client)
@@ -26,6 +27,8 @@ namespace Library.Amazon
 
         public async Task<bool> StorageExistsAsync(string storage, CancellationToken token = default)
         {
+            if (storage == null) throw new ArgumentNullException(nameof(storage));
+
             // TODO: catch exception and throw one similar to exceptions thrown in other methods?
             var exists = await _client.DoesS3BucketExistAsync(storage);
             return exists;
@@ -33,42 +36,54 @@ namespace Library.Amazon
 
         public async Task CreateStorageAsync(string storage, CancellationToken token = default)
         {
-            var mutex = new Mutex(true, $"AWS:bucket:{storage}");
+            if (storage == null) throw new ArgumentNullException(nameof(storage));
+
+            var entered = await _createSemaphore.WaitAsync(5000, token);
+
+            // TODO: Find proper exception to throw and add to string resources
+            if (!entered) throw new Exception("Bucket creation aborted to avoid deadlock in critical section of code.");
 
             try
             {
-                mutex.WaitOne(5000);
-
-                // TODO: call _client.DoesS3BucketExistAsync(storage) instead, and
                 // TODO: catch exception and throw one similar to exceptions thrown in other methods
-                var exists = await StorageExistsAsync(storage, token);
+                var exists = await _client.DoesS3BucketExistAsync(storage);
                 if (exists) return;
-             
-                var response = await _client.PutBucketAsync(storage, token);
-                ExceptionHelper.ThrowOnFailedHttpRequest(response.HttpStatusCode, "S3StorageManager:CreateStorageAsync:HttpRequestException", storage);
+
+                var request = new PutBucketRequest
+                {
+                    BucketName = storage,
+                    UseClientRegion = false
+                };
+
+                var response = await _client.PutBucketAsync(request, token);
+                Resources.ExceptionHelper.ThrowOnFailedHttpRequest(response.HttpStatusCode, "S3StorageManager:CreateStorageAsync:HttpRequestException", storage);
 
                 // Give AWS time to guarantee bucket creation
                 await Task.Delay(1200, token);
             }
             finally
             {
-                mutex.ReleaseMutex();
+                _createSemaphore.Release();
             }
         }
 
         public async Task DeleteStorageAsync(string storage, CancellationToken token = default)
         {
+            if (storage == null) throw new ArgumentNullException(nameof(storage));
+
             var response = await _client.DeleteBucketAsync(storage, token);
-            ExceptionHelper.ThrowOnFailedHttpRequest(response.HttpStatusCode, "S3StorageManager:DeleteStorageAsync:HttpRequestException", storage);
+            Resources.ExceptionHelper.ThrowOnFailedHttpRequest(response.HttpStatusCode, "S3StorageManager:DeleteStorageAsync:HttpRequestException", storage);
         }
 
         public async Task PurgeStorageAsync(string storage, CancellationToken token = default)
         {
+            if (storage == null) throw new ArgumentNullException(nameof(storage));
+
             var block = new ActionBlock<string>(async key =>
             {
                 var deleteRequest = new DeleteObjectRequest {BucketName = storage, Key = key};
                 var deleteResponse = await _client.DeleteObjectAsync(deleteRequest, token);
-                ExceptionHelper.ThrowOnFailedHttpRequest(deleteResponse.HttpStatusCode, "S3StorageManager:PurgeStorageAsync_DeleteObject:HttpRequestException", storage);
+                Resources.ExceptionHelper.ThrowOnFailedHttpRequest(deleteResponse.HttpStatusCode, "S3StorageManager:PurgeStorageAsync_DeleteObject:HttpRequestException", storage);
             });
 
             var listRequest = new ListObjectsV2Request {BucketName = storage, Prefix = String.Empty};
@@ -77,7 +92,7 @@ namespace Library.Amazon
             do
             {
                 listResponse = await _client.ListObjectsV2Async(listRequest, token);
-                ExceptionHelper.ThrowOnFailedHttpRequest(listResponse.HttpStatusCode, "S3StorageManager:PurgeStorageAsync_ListObjects:HttpRequestException", storage);
+                Resources.ExceptionHelper.ThrowOnFailedHttpRequest(listResponse.HttpStatusCode, "S3StorageManager:PurgeStorageAsync_ListObjects:HttpRequestException", storage);
 
                 foreach (var s3Object in listResponse.S3Objects)
                 {
@@ -94,7 +109,7 @@ namespace Library.Amazon
         public async IAsyncEnumerable<string> ListStoragesAsync([EnumeratorCancellation] CancellationToken token = default)
         {
             var response = await _client.ListBucketsAsync(token);
-            ExceptionHelper.ThrowOnFailedHttpRequest(response.HttpStatusCode, "S3StorageManager:ListStoragesAsync:HttpRequestException");
+            Resources.ExceptionHelper.ThrowOnFailedHttpRequest(response.HttpStatusCode, "S3StorageManager:ListStoragesAsync:HttpRequestException");
 
             foreach (var bucket in response.Buckets)
             {
