+++ src/Library.Amazon/SqsQueueManager.cs
@@ -7,16 +7,17 @@ using System.Runtime.CompilerServices;
 using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
+using Amazon.S3.Model.Internal.MarshallTransformations;
 using Amazon.SQS;
 using Amazon.SQS.Model;
-using Library.Http;
-using Library.Queuing;
 using Library.Amazon.Resources;
+using Library.Platform.Queuing;
 
 namespace Library.Amazon
 {
     public class SqsQueueManager : IQueueManager
     {
+        private static readonly SemaphoreSlim _createSemaphore = new SemaphoreSlim(1, 1);
         private readonly IAmazonSQS _client;
 
         public SqsQueueManager(IAmazonSQS client)
@@ -27,31 +28,28 @@ namespace Library.Amazon
         
         public async Task<bool> QueueExistsAsync(string queue, CancellationToken token = default)
         {
-            var response = await _client.GetQueueUrlAsync(queue, token);
-            if (response.HttpStatusCode == HttpStatusCode.NotFound) return false;
-            ExceptionHelper.ThrowOnFailedHttpRequest(response.HttpStatusCode, "SqsQueueManager:QueueExistsAsync:HttpRequestException", queue);
+            if (queue == null) throw new ArgumentNullException(nameof(queue));
 
-            return true;
+            var response = await _client.ListQueuesAsync(queue, token);
+            Resources.ExceptionHelper.ThrowOnFailedHttpRequest(response.HttpStatusCode, "SqsQueueManager:QueueExistsAsync:HttpRequestException", queue);
+
+            var exists = response.QueueUrls.Any(url => url[(url.LastIndexOf('/') + 1)..].Equals(queue, StringComparison.OrdinalIgnoreCase));
+            return exists;
         }
 
         public async Task<string> CreateQueueAsync(string queue, CancellationToken token = default)
         {
-            var mutex = new Mutex(true, $"AWS:queue:{queue}");
+            if (queue == null) throw new ArgumentNullException(nameof(queue));
+
+            var entered = await _createSemaphore.WaitAsync(5000, token);
+
+            // TODO: Find proper exception to throw and add to string resources
+            if (!entered) throw new Exception("Queue creation aborted to avoid deadlock in critical section of code.");
 
             try
             {
-                mutex.WaitOne(5000);
-
-                // TODO: Optimize this. QueueExistsAsync internally calls GetQueueUrlAsync
-                var exists = await QueueExistsAsync(queue, token);
-                if (exists)
-                {
-                    // TODO: don't call this method, but perform the steps here
-                    return await GetQueueUrlAsync(queue, token);
-                }
-
                 var response = await _client.CreateQueueAsync(queue, token);
-                ExceptionHelper.ThrowOnFailedHttpRequest(response.HttpStatusCode, "SqsQueueManager:CreateQueueAsync:HttpRequestException", queue);
+                Resources.ExceptionHelper.ThrowOnFailedHttpRequest(response.HttpStatusCode, "SqsQueueManager:CreateQueueAsync:HttpRequestException", queue);
 
                 // Give AWS time to guarantee queue creation
                 await Task.Delay(1200, token);
@@ -60,30 +58,42 @@ namespace Library.Amazon
             }
             finally
             {
-                mutex.ReleaseMutex();
+                _createSemaphore.Release();
             }
         }
         
         public async Task DeleteQueueAsync(string queue, CancellationToken token = default)
         {
-            // TODO: don't call this method, but perform the steps here
-            var url = await GetQueueUrlAsync(queue, token);
-            var response = await _client.DeleteQueueAsync(url, token);
-            ExceptionHelper.ThrowOnFailedHttpRequest(response.HttpStatusCode, "SqsQueueManager:DeleteQueueAsync:HttpRequestException", queue);
+            if (queue == null) throw new ArgumentNullException(nameof(queue));
+
+            var listResponse = await _client.ListQueuesAsync(queue, token);
+            Resources.ExceptionHelper.ThrowOnFailedHttpRequest(listResponse.HttpStatusCode, "SqsQueueManager:QueueExistsAsync:HttpRequestException", queue);
+
+            var url = listResponse.QueueUrls.SingleOrDefault(url => url[(url.LastIndexOf('/') + 1)..].Equals(queue, StringComparison.OrdinalIgnoreCase));
+            if (url == null) return;
+
+            var deleteResponse = await _client.DeleteQueueAsync(url, token);
+            Resources.ExceptionHelper.ThrowOnFailedHttpRequest(deleteResponse.HttpStatusCode, "SqsQueueManager:DeleteQueueAsync:HttpRequestException", queue);
         }
         
         public async Task PurgeQueueAsync(string queue, CancellationToken token = default)
         {
-            // TODO: don't call this method, but perform the steps here
-            var url = await GetQueueUrlAsync(queue, token);
-            var response = await _client.PurgeQueueAsync(url, token);
-            ExceptionHelper.ThrowOnFailedHttpRequest(response.HttpStatusCode, "SqsQueueManager:PurgeQueueAsync:HttpRequestException", queue);
+            if (queue == null) throw new ArgumentNullException(nameof(queue));
+
+            var listResponse = await _client.ListQueuesAsync(queue, token);
+            Resources.ExceptionHelper.ThrowOnFailedHttpRequest(listResponse.HttpStatusCode, "SqsQueueManager:QueueExistsAsync:HttpRequestException", queue);
+
+            var url = listResponse.QueueUrls.SingleOrDefault(u => u[(u.LastIndexOf('/') + 1)..].Equals(queue, StringComparison.OrdinalIgnoreCase));
+            if (url == null) return;
+
+            var purgeResponse = await _client.PurgeQueueAsync(url, token);
+            Resources.ExceptionHelper.ThrowOnFailedHttpRequest(purgeResponse.HttpStatusCode, "SqsQueueManager:PurgeQueueAsync:HttpRequestException", queue);
         }
         
         public async IAsyncEnumerable<string> ListQueuesAsync([EnumeratorCancellation] CancellationToken token = default)
         {
             var response = await _client.ListQueuesAsync(String.Empty, token);
-            ExceptionHelper.ThrowOnFailedHttpRequest(response.HttpStatusCode, "SqsQueueManager:ListQueuesAsync:HttpRequestException");
+            Resources.ExceptionHelper.ThrowOnFailedHttpRequest(response.HttpStatusCode, "SqsQueueManager:ListQueuesAsync:HttpRequestException");
 
             foreach (var url in response.QueueUrls)
             {
@@ -94,10 +104,13 @@ namespace Library.Amazon
 
         public async Task<string> GetQueueUrlAsync(string queue, CancellationToken token = default)
         {
-            var response = await _client.GetQueueUrlAsync(queue, token);
-            ExceptionHelper.ThrowOnFailedHttpRequest(response.HttpStatusCode, "SqsQueueManager:GetQueueUrlAsync:HttpRequestException", queue);
+            if (queue == null) throw new ArgumentNullException(nameof(queue));
 
-            return response.QueueUrl;
+            var listResponse = await _client.ListQueuesAsync(queue, token);
+            Resources.ExceptionHelper.ThrowOnFailedHttpRequest(listResponse.HttpStatusCode, "SqsQueueManager:GetQueueUrlAsync:HttpRequestException", queue);
+
+            var url = listResponse.QueueUrls.SingleOrDefault(url => url[(url.LastIndexOf('/') + 1)..].Equals(queue, StringComparison.OrdinalIgnoreCase));
+            return url;
         }
     }
 }
